fix: add code review changes
